more simplifications
