+++ src/test/gap_visibility.svelte.test.js
@@ -20,114 +20,17 @@
 
 import { describe, it, expect, beforeEach } from 'vitest';
 import { render } from 'vitest-browser-svelte';
-import { tick } from 'svelte';
-import Session from '../lib/Session.svelte.js';
 import SveditTest from './testing_components/SveditTest.svelte';
-import { document_schema, session_config } from '../routes/create_demo_session.js';
-import nanoid from '../routes/nanoid.js';
-
-/** Sleep one or more requestAnimationFrame ticks. */
-function raf(times = 1) {
-	return new Promise((resolve) => {
-		let remaining = times;
-		function step() {
-			remaining -= 1;
-			if (remaining <= 0) resolve();
-			else requestAnimationFrame(step);
-		}
-		requestAnimationFrame(step);
-	});
-}
-
-/** Let Svelte + IO + scroll-listener RAFs settle. */
-async function settle() {
-	await tick();
-	await raf(3);
-	await new Promise((r) => setTimeout(r, 50));
-}
-
-/** Build a session with a single story whose `buttons` array holds `n` buttons. */
-function make_story_session(n_buttons) {
-	const button_ids = [];
-	const nodes = {};
-	for (let i = 0; i < n_buttons; i++) {
-		const id = `btn_${i}`;
-		nodes[id] = {
-			id,
-			type: 'button',
-			label: { text: `Action ${i + 1}`, annotations: [] },
-			href: '#'
-		};
-		button_ids.push(id);
-	}
-	nodes.story_1 = {
-		id: 'story_1',
-		type: 'story',
-		layout: 1,
-		image: '',
-		title: { text: 'Test story', annotations: [] },
-		description: { text: 'desc', annotations: [] },
-		buttons: button_ids
-	};
-	nodes.page_1 = {
-		id: 'page_1',
-		type: 'page',
-		body: ['story_1'],
-		keywords: [],
-		daily_visitors: [],
-		created_at: '2025-05-30T10:39:59.987Z'
-	};
-	return new Session(document_schema, { document_id: 'page_1', nodes }, { ...session_config });
-}
-
-/** Build a session with a single image_grid containing `n` items. */
-function make_image_grid_session(n_items) {
-	const item_ids = [];
-	const nodes = {};
-	for (let i = 0; i < n_items; i++) {
-		const id = `igi_${i}`;
-		nodes[id] = {
-			id,
-			type: 'image_grid_item',
-			image: '',
-			title: { text: `Item ${i + 1}`, annotations: [] },
-			description: { text: 'desc', annotations: [] }
-		};
-		item_ids.push(id);
-	}
-	nodes.image_grid_1 = {
-		id: 'image_grid_1',
-		type: 'image_grid',
-		layout: 1,
-		image_grid_items: item_ids
-	};
-	nodes.page_1 = {
-		id: 'page_1',
-		type: 'page',
-		body: ['image_grid_1'],
-		keywords: [],
-		daily_visitors: [],
-		created_at: '2025-05-30T10:39:59.987Z'
-	};
-	return new Session(document_schema, { document_id: 'page_1', nodes }, { ...session_config });
-}
-
-/** Find the buttons array element inside the rendered tree. */
-function find_buttons_array(container) {
-	return container.querySelector('[data-type="node_array"][data-path$="buttons"]:not(.empty)');
-}
-
-function find_image_grid_array(container) {
-	return container.querySelector('[data-type="node_array"][data-path$="image_grid_items"]');
-}
-
-function find_last_gap(array_el) {
-	return array_el.querySelector(':scope > .node-gap.gap-after.last');
-}
-
-function find_first_gap(array_el) {
-	return array_el.querySelector(':scope > .node-gap.gap-before:not(.empty)');
-}
+import {
+	settle,
+	settle_grid,
+	make_story_session,
+	make_image_grid_session,
+	find_buttons_array,
+	find_image_grid_array,
+	find_last_gap,
+	find_first_gap
+} from './test_utils.js';
 
 describe('NodeGap visibility & placement', () => {
 	beforeEach(() => {
@@ -248,16 +151,6 @@ describe('NodeGap visibility & placement', () => {
 	});
 
 	describe('wrap layout (image grid)', () => {
-		// image_grid_items render with aspect-ratio: 1/1 width: 100%, so
-		// the grid layout settles a frame or two later than a row of fixed-
-		// width buttons. Give the IO + edge_map sync extra time before
-		// asserting .positioned.
-		async function settle_grid() {
-			await tick();
-			await raf(8);
-			await new Promise((r) => setTimeout(r, 150));
-		}
-
 		// Helper: scroll the LAST item itself into the IO overscan zone
 		// and wait for the IO to populate near_map. Centering the whole
 		// array doesn't work when the array is taller than the viewport
